Add project files and improve  project structure.
For python projects it is recommended to use
a tool like *Poetry* to manage dependencies easily.

It allows you to tell user which dependencies
are necessary, specify versions...
It is a great tool for python and easy to learn
following the official doc.

I also added .vscode to .gitignore because
it is a folder which is different for each
programmer, so it should not be pushed